tui_protocol/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rootlessnet-tui"
version = "1.0.0"
description = "RootlessNet TUI Protocol - Decentralized Tor Blockchain Network"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "RootlessNet Community", email = "aaryan.bansal.dev@gmail.com"}
]
keywords = ["tor", "blockchain", "decentralized", "tui", "privacy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications",
    "Topic :: Internet",
    "Topic :: Security :: Cryptography",
]

dependencies = [
    "textual>=0.40.0",
    "rich>=13.0.0",
    "cryptography>=42.0.4",
    "aiofiles>=23.0.0",
    "pynacl>=1.5.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
tor = ["aiohttp-socks>=0.8.0", "stem>=1.8.0"]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0", "mypy>=1.0.0"]

[project.scripts]
rootlessnet = "rootlessnet.main:main"

[project.urls]
Homepage = "https://github.com/RootlessNet/protocol"
Documentation = "https://github.com/RootlessNet/protocol#readme"
Repository = "https://github.com/RootlessNet/protocol"

[tool.setuptools.packages.find]
where = ["."]
